Split Randomizer into smaller interfaces
#150
